@jinyus I saw C++ is using `-O3`, while I think we should just not use `-O3` for any language, it's your call in the end.

Meanwhile, can you kindly try `-O3` on the VM and see if it's faster or slower? I honestly don't know what to expect at this point, won't be surprised if it's actually slower, Julia's `O2` is basically always as fast.