@michaelsbradleyjr 2023-10-29
@Moelf fwiw, at one point when working on the `nim_con` implementation, I got very frustrated with my local results (on mac laptop, or Docker container in VM on that laptop) being a mismatch for the official results.

I then signed up for an Azure account and automatically received free credits that equate to about a month of running a `Standard F4s v2 (4 vcpus, 8 GiB memory)` VM. Choosing the Ubuntu 22.04 image and then installing Docker on it, it's the same setup used for the official benchmarking.

With that setup, I gained a lot of confidence for trying various changes to my code because I know how it will turn out when jinyus runs it.