@jinyus 2023-10-29
Main:
`Julia | 28.31 ms | 337.00 ms | 2.92 s`

This PR:
```
Julia:

        Benchmark 1: julia -O3 --project=Related -e "using Related; main()"
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 22 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 24 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
        Processing time (w/o IO): 23 milliseconds
          Time (mean ± σ):      3.357 s ±  0.026 s    [User: 3.208 s, System: 0.253 s]
          Range (min … max):    3.321 s …  3.393 s    10 runs
         
Julia:

        Benchmark 1: julia -O3 --project=Related -e "using Related; main()"
        Processing time (w/o IO): 336 milliseconds
        Processing time (w/o IO): 338 milliseconds
        Processing time (w/o IO): 337 milliseconds
          Time (mean ± σ):      4.821 s ±  0.010 s    [User: 4.612 s, System: 0.311 s]
          Range (min … max):    4.815 s …  4.828 s    2 runs
         
Julia:

        Benchmark 1: julia -O3 --project=Related -e "using Related; main()"
        Processing time (w/o IO): 2913 milliseconds
        Processing time (w/o IO): 2912 milliseconds
        Processing time (w/o IO): 2910 milliseconds
          Time (mean ± σ):     16.374 s ±  0.019 s    [User: 16.112 s, System: 0.356 s]
          Range (min … max):   16.360 s … 16.387 s    2 runs

```