@Moelf 2023-10-29
Wait is that faster lol ... Interesting