@Moelf 2023-10-30
@michaelsbradleyjr 
![image](https://github.com/jinyus/related_post_gen/assets/5306213/0a13a60e-c265-49f7-b194-1f76de8e86f2)

ehh, how much free credit did you get?