@Tortar 2023-10-30
@Moelf can you try with the biggest benchmark? I see a lot of variance on the small benchmarks, maybe trying with the bigger ones can lead to more consistent results (I always used the smallest so I'm not sure)