@Tortar 2023-10-30
Probably the behaviours is the same as the one explained here though: https://discourse.julialang.org/t/why-does-my-programs-performance-vary-so-much-from-run-to-run-can-it-be-fixed/53678